% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scImmuCC.R
\name{scImmuCC}
\alias{scImmuCC}
\title{scImmuCC}
\usage{
scImmuCC(
  count,
  Non_Immune = TRUE,
  resultsdir = "",
  genelist_dir = "data",
  min_cells = 50,
  ncores = 1
)
}
\arguments{
\item{count}{a matrix with cell unique barcodes as column names and gene names as row names .}

\item{Non_Immune}{Whether non-immune cells are included in the matrix.}
}
\value{
Data frames with the barcodes and cell types, and some maps.
}
\description{
Creating Hierarchical_annotation for scRNA-Seq data immune cell.
}
\details{
Input takes a cells-genes matrix with cell unique barcodes as column names and gene names as row names and returns the cells annotation.
}
\examples{
test_data
library(GSVA)
library(Seurat)
}
