% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-ImmuCC.R
\name{scImmuCC_Layered}
\alias{scImmuCC_Layered}
\title{scImmuCC_Layered.}
\usage{
scImmuCC_Layered(count, Non_Immune = TRUE)
}
\arguments{
\item{count}{a matrix with cell unique barcodes as column names and gene names as row names .}

\item{Non_Immune}{Whether non-immune cells are included in the matrix.}
}
\value{
Data frames with the barcodes and cell types, and some maps.
}
\description{
Creating Hierarchical_annotation for scRNA-Seq data immune cell.
}
\details{
Input takes a cells-genes matrix with cell unique barcodes as column names and gene names as row names and returns the cells annotation.
}
\examples{
test_data
}
