% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scImmuCC_main.R
\name{scImmuCC_main}
\alias{scImmuCC_main}
\title{scImmuCC_main.}
\usage{
scImmuCC_main(count, genematrix, filename)
}
\arguments{
\item{count}{a matrix with cell unique barcodes as column names and gene names as row names .}

\item{genematrix}{a data frame with cell types as column names .}

\item{filename}{custom file name， character .}

\item{ssGSEA_result}{a data frame , scImmuCC return result .}
}
\value{
a data frame with cell annotation.
}
\description{
Calculation of gene enrichment scores and annotate cell type .
}
\details{
Input takes a count, genelist and filename, returns a data frame with cell annotation .
}
