% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_maps.R
\name{seurat_Heatmap}
\alias{seurat_Heatmap}
\title{Seurat Object.}
\usage{
seurat_Heatmap(count, genematrix, ssGSEA_result, filename)
}
\arguments{
\item{count}{a matrix with cell unique barcodes as column names and gene names as row names .}

\item{genematrix}{a data frame with cell types as column names .}

\item{ssGSEA_result}{a data frame , scImmuCC return result .}

\item{filename}{custom file name， character .}
}
\value{
4 pictures.
}
\description{
Creating dynamic Seurat Object to add annotation information and draw images.
}
\details{
Input takes a count, genelist and scImmuCC result, returns tsne, umap, Dotplot and pheatmaps.
}
